Revert "Build libmedia_native module in order to get mediaserver running"
This reverts commit 536683a29205073d406f14380676dd030ad642f7.

I've been trolled! I didn't think the Pyramid was using a prebuilt libcameraservice.so, so when I saw that libcameraservice.so depended on libmedia_native.so I thought we needed to add libmedia_native.so. It turns out that the libcameraservice.so file was built for CM10.1 so at the time it was built it was built to depend on libmedia_native.so which is deprecated in Android 4.3/CM10.2. Libmedia_native.so is just libmedia.so split into an extra lib.